on inexact matching of notes
Assuming the untimed phrase: 'C4,D4,E4,D4' is an inexact attempt, the first example (replace) seems to be correct if the original matching code was 'D4,D4,E4,D4'. 

However, on the second example, if the user has played 'C4, B4, D4', would not this mean that they have made both a replace and an insert error? (C4 replacing D4 and B4 as an extra insert, i.e. worth 3). 
Would 'D4,B4,D4' be the case where B4 is the insert on the original code?

The third example seems to be very similar as the second one; would 'D4,/1,E4,D4' be the case where the second 'D4' is omitted (hence the delay)?